+++ codegen/table_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/cloudquery/plugin-sdk/schema"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/go-cmp/cmp/cmpopts"
+	"github.com/stretchr/testify/require"
 )
 
 type (
@@ -77,6 +78,7 @@ var (
 			Name:     "int_col",
 			Type:     schema.TypeInt,
 			Resolver: `schema.PathResolver("IntCol")`,
+			Options:  schema.ColumnCreationOptions{PrimaryKey: true},
 		},
 		{
 			Name:     "int64_col",
@@ -154,7 +156,12 @@ var (
 			// Should not be unwrapped
 			ColumnDefinition{Name: "test_struct", Type: schema.TypeJSON, Resolver: `schema.PathResolver("TestStruct")`},
 			// Should be unwrapped
-			ColumnDefinition{Name: "non_embedded_embedded_string", Type: schema.TypeString, Resolver: `schema.PathResolver("NonEmbedded.EmbeddedString")`},
+			ColumnDefinition{
+				Name:     "non_embedded_embedded_string",
+				Type:     schema.TypeString,
+				Resolver: `schema.PathResolver("NonEmbedded.EmbeddedString")`,
+				Options:  schema.ColumnCreationOptions{PrimaryKey: true},
+			},
 		},
 		nameTransformer:     DefaultNameTransformer,
 		typeTransformer:     DefaultTypeTransformer,
@@ -167,6 +174,7 @@ var (
 				Name:     "time_col",
 				Type:     schema.TypeTimestamp,
 				Resolver: `schema.PathResolver("TimeCol")`,
+				Options:  schema.ColumnCreationOptions{PrimaryKey: true},
 			},
 			{
 				Name:     "custom",
@@ -227,6 +235,7 @@ func TestTableFromGoStruct(t *testing.T) {
 			name: "should generate table from struct with default options",
 			args: args{
 				testStruct: testStruct{},
+				options:    []TableOption{WithPKColumns("int_col")},
 			},
 			want: expectedTestTable,
 		},
@@ -234,7 +243,10 @@ func TestTableFromGoStruct(t *testing.T) {
 			name: "should unwrap all embedded structs when option is set",
 			args: args{
 				testStruct: testStructWithEmbeddedStruct{},
-				options:    []TableOption{WithUnwrapAllEmbeddedStructs()},
+				options: []TableOption{
+					WithPKColumns("int_col"),
+					WithUnwrapAllEmbeddedStructs(),
+				},
 			},
 			want: expectedTestTableEmbeddedStruct,
 		},
@@ -242,7 +254,10 @@ func TestTableFromGoStruct(t *testing.T) {
 			name: "should unwrap specific structs when option is set",
 			args: args{
 				testStruct: testStructWithNonEmbeddedStruct{},
-				options:    []TableOption{WithUnwrapStructFields([]string{"NonEmbedded"})},
+				options: []TableOption{
+					WithPKColumns("non_embedded_embedded_string"),
+					WithUnwrapStructFields([]string{"NonEmbedded"}),
+				},
 			},
 			want: expectedTestTableNonEmbeddedStruct,
 		},
@@ -250,30 +265,46 @@ func TestTableFromGoStruct(t *testing.T) {
 			name: "should override schema type when option is set",
 			args: args{
 				testStruct: testStructWithCustomType{},
-				options: []TableOption{WithTypeTransformer(func(t reflect.StructField) (schema.ValueType, error) {
-					switch t.Type {
-					case reflect.TypeOf(time.Time{}), reflect.TypeOf(&time.Time{}):
-						return schema.TypeJSON, nil
-					default:
-						return schema.TypeInvalid, nil
-					}
-				})},
+				options: []TableOption{
+					WithPKColumns("time_col"),
+					WithTypeTransformer(func(t reflect.StructField) (schema.ValueType, error) {
+						switch t.Type {
+						case reflect.TypeOf(time.Time{}), reflect.TypeOf(&time.Time{}):
+							return schema.TypeJSON, nil
+						default:
+							return schema.TypeInvalid, nil
+						}
+					}),
+				},
 			},
 			want: TableDefinition{Name: "test_struct",
 				// We expect the time column to be of type JSON, since we override the type of `time.Time` to be JSON
-				Columns:         ColumnDefinitions{{Name: "time_col", Type: schema.TypeJSON, Resolver: `schema.PathResolver("TimeCol")`}},
+				Columns: ColumnDefinitions{{
+					Name:     "time_col",
+					Type:     schema.TypeJSON,
+					Resolver: `schema.PathResolver("TimeCol")`,
+					Options:  schema.ColumnCreationOptions{PrimaryKey: true},
+				}},
 				nameTransformer: DefaultNameTransformer},
 		},
 		{
 			name: "should handle default and custom acronyms correctly",
 			args: args{
 				testStruct: testStructCaseCheck{},
-				options:    []TableOption{WithNameTransformer(customNameTransformer)},
+				options: []TableOption{
+					WithPKColumns("ip_address"),
+					WithNameTransformer(customNameTransformer),
+				},
 			},
 			want: TableDefinition{Name: "test_struct",
 				// We expect the time column to be of type JSON, since we override the type of `time.Time` to be JSON
 				Columns: ColumnDefinitions{
-					{Name: "ip_address", Type: schema.TypeString, Resolver: `schema.PathResolver("IPAddress")`},
+					{
+						Name:     "ip_address",
+						Type:     schema.TypeString,
+						Resolver: `schema.PathResolver("IPAddress")`,
+						Options:  schema.ColumnCreationOptions{PrimaryKey: true},
+					},
 					{Name: "cdns", Type: schema.TypeString, Resolver: `schema.PathResolver("CDNs")`},
 					{Name: "my_cdn", Type: schema.TypeString, Resolver: `schema.PathResolver("MyCDN")`},
 					{Name: "cidr", Type: schema.TypeInt, Resolver: `schema.PathResolver("CIDR")`},
@@ -291,6 +322,7 @@ func TestTableFromGoStruct(t *testing.T) {
 			args: args{
 				testStruct: testStructForCustomResolvers{},
 				options: []TableOption{
+					WithPKColumns("time_col"),
 					WithTypeTransformer(customTypeTransformer),
 					WithResolverTransformer(customResolverTransformer),
 				},
@@ -305,7 +337,7 @@ func TestTableFromGoStruct(t *testing.T) {
 			if err != nil {
 				t.Fatal(err)
 			}
-			if diff := cmp.Diff(table, &tt.want,
+			if diff := cmp.Diff(&tt.want, table,
 				cmpopts.IgnoreUnexported(TableDefinition{})); diff != "" {
 				t.Fatalf("table does not match expected. diff (-got, +want): %v", diff)
 			}
@@ -314,6 +346,53 @@ func TestTableFromGoStruct(t *testing.T) {
 				t.Fatal(err)
 			}
 			fmt.Println(buf.String())
+		})
+	}
+}
+
+func TestTableDefinition_Check(t *testing.T) {
+	tests := []struct {
+		name       string
+		definition *TableDefinition
+		err        error
+	}{
+		{
+			name: "nil",
+			err:  fmt.Errorf("nil table definition"),
+		},
+		{
+			name:       "no table name",
+			definition: new(TableDefinition),
+			err:        fmt.Errorf("empty table name"),
+		},
+		{
+			name:       "no columns",
+			definition: &TableDefinition{Name: "no_columns"},
+			err:        fmt.Errorf("no columns for table no_columns"),
+		},
+		{
+			name: "invalid type",
+			definition: &TableDefinition{
+				Name:    "invalid_type",
+				Columns: ColumnDefinitions{{Name: "col1"}},
+			},
+			err: fmt.Errorf("invalid_type->col1: invalid column type"),
+		},
+		{
+			name: "duplicate name",
+			definition: &TableDefinition{
+				Name: "duplicate_name",
+				Columns: ColumnDefinitions{
+					{Name: "col1", Type: schema.TypeInt},
+					{Name: "col1", Type: schema.TypeInt},
+				},
+			},
+			err: fmt.Errorf("duplicate_name->col1: duplicate column name"),
+		},
+	}
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			require.EqualValues(t, tt.err, tt.definition.Check())
 		})
 	}
 }